+++ docs/source/changelog.rst
@@ -3,11 +3,14 @@
 ChangeLog
 #########
 
+=======
 Version 0.123.1
 ===============
 
 * Add backward compatibility to deserialize OpenRAVE::geometry::RaveOrientedBox<T>
 
+* Add "axial" geometry type.
+    
 Version 0.123.0
 ===============